+++ blackmanHarris.js
@@ -1,0 +1,2 @@
+import { cosineSum } from './_util.js'
+export function blackmanHarris (i, N) { return cosineSum(i, N, [0.35875, 0.48829, 0.14128, 0.01168]) }
