+++ blackmanNuttall.js
@@ -1,0 +1,2 @@
+import { cosineSum } from './_util.js'
+export function blackmanNuttall (i, N) { return cosineSum(i, N, [0.3635819, 0.4891775, 0.1365995, 0.0106411]) }
