+++ connes.js
@@ -1,0 +1,1 @@
+export function connes (i, N) { let x = (2 * i - N + 1) / (N - 1); return (1 - x * x) * (1 - x * x) }
