+++ cosine.js
@@ -1,0 +1,2 @@
+import { sin, PI } from './_util.js'
+export function cosine (i, N) { return sin(PI * i / (N - 1)) }
