+++ flatTop.js
@@ -1,0 +1,2 @@
+import { cosineSum } from './_util.js'
+export function flatTop (i, N) { return cosineSum(i, N, [1, 1.93, 1.29, 0.388, 0.028]) }
