+++ generate.js
@@ -1,0 +1,5 @@
+export function generate (fn, N, ...params) {
+	let w = new Float64Array(N)
+	for (let i = 0; i < N; i++) w[i] = fn(i, N, ...params)
+	return w
+}
