+++ nuttall.js
@@ -1,0 +1,2 @@
+import { cosineSum } from './_util.js'
+export function nuttall (i, N) { return cosineSum(i, N, [0.355768, 0.487396, 0.144232, 0.012604]) }
