+++ package.json
@@ -3,8 +3,50 @@
   "version": "3.0.0",
   "type": "module",
   "description": "Window functions for signal processing and spectral analysis",
-  "exports": "./index.js",
+  "exports": {
+    ".": "./index.js",
+    "./rectangular": "./rectangular.js",
+    "./triangular": "./triangular.js",
+    "./bartlett": "./bartlett.js",
+    "./welch": "./welch.js",
+    "./connes": "./connes.js",
+    "./hann": "./hann.js",
+    "./hamming": "./hamming.js",
+    "./cosine": "./cosine.js",
+    "./blackman": "./blackman.js",
+    "./exactBlackman": "./exactBlackman.js",
+    "./nuttall": "./nuttall.js",
+    "./blackmanNuttall": "./blackmanNuttall.js",
+    "./blackmanHarris": "./blackmanHarris.js",
+    "./flatTop": "./flatTop.js",
+    "./bartlettHann": "./bartlettHann.js",
+    "./lanczos": "./lanczos.js",
+    "./parzen": "./parzen.js",
+    "./bohman": "./bohman.js",
+    "./powerOfSine": "./powerOfSine.js",
+    "./kaiser": "./kaiser.js",
+    "./gaussian": "./gaussian.js",
+    "./generalizedNormal": "./generalizedNormal.js",
+    "./tukey": "./tukey.js",
+    "./planckTaper": "./planckTaper.js",
+    "./exponential": "./exponential.js",
+    "./hannPoisson": "./hannPoisson.js",
+    "./cauchy": "./cauchy.js",
+    "./rifeVincent": "./rifeVincent.js",
+    "./confinedGaussian": "./confinedGaussian.js",
+    "./kaiserBesselDerived": "./kaiserBesselDerived.js",
+    "./dolphChebyshev": "./dolphChebyshev.js",
+    "./taylor": "./taylor.js",
+    "./dpss": "./dpss.js",
+    "./ultraspherical": "./ultraspherical.js",
+    "./generate": "./generate.js",
+    "./apply": "./apply.js",
+    "./enbw": "./enbw.js",
+    "./scallopLoss": "./scallopLoss.js",
+    "./cola": "./cola.js"
+  },
   "types": "./index.d.ts",
+  "typesVersions": { "*": { "*": ["./index.d.ts"] } },
   "scripts": {
     "test": "node --test test.js"
   },
