+++ powerOfSine.js
@@ -1,0 +1,5 @@
+import { sin, PI, pow } from './_util.js'
+export function powerOfSine (i, N, alpha) {
+	if (alpha == null) alpha = 2
+	return pow(sin(PI * i / (N - 1)), alpha)
+}
