+++ rectangular.js
@@ -1,0 +1,1 @@
+export function rectangular () { return 1 }
