+++ welch.js
@@ -1,0 +1,1 @@
+export function welch (i, N) { let x = (2 * i - N + 1) / (N - 1); return 1 - x * x }
